Fixes #255 - organizations can now accept registrations for dances and lessons without users having to have an account - however members will still need to have accounts so that the individual organizations can track membership
