+++ app/views/organization_orders/_payment_button.html.slim
@@ -3,7 +3,7 @@
 /- if [Payable::Methods::STRIPE, Payable::Methods::PAYPAL].include?(order.payment_method)
 - if event.has_payment_processor?(Integration::STRIPE)
   = form_tag(payments_path(host_id: attendance.host.id, host_type: attendance.host.class.name, order_id: order.id)) do
-    = stripe_checkout_script(event.name, order.total)
+    = stripe_checkout_script(event.name, order.total, email: attendance.attendee_email)
 
 span.small
   span<> Paying with:
